[🐛fix]: #62 마이페이지 드롭다운 누락 수정
