+++ src/api/authApi.js
@@ -45,7 +45,7 @@ export const userSignOut = async (signInData) => {
 //회원가입 -> 이메일 인증
 export const userSignUpEmail = async (email) => {
   const url = '/auth/sign-up/email';
-  const response = await api.post(url, email);
+  const response = await api.post(url, { email: email });
   return response;
 };
